+++ CHANGELOG.md
@@ -46,6 +46,8 @@ Release: dd.mm.yyyy
 
 - Change: Function `mainScreen()` is now simply `main()` ([#108](https://github.com/kasper/phoenix/issues/108)).
 - Change: Function `screens()` is now `all()` ([#108](https://github.com/kasper/phoenix/issues/108)).
+- Change: Function `windows()` now takes optionals `windows(Map<String, AnyObject> optionals)` ([#108](https://github.com/kasper/phoenix/issues/108)).
+- Deprecation: Function `visibleWindows()` has been removed, use `windows({ visible: true })` instead ([#108](https://github.com/kasper/phoenix/issues/108)).
 
 #### Command